pyproject.toml

[project]
name = "mqttgpio"
version = "0.0.2"
description = "MQTT GPIO Controller for Raspberry Pi"
authors = [{ name = "James Hodgkinson", email = "james@terminaloutcomes.com" }]
license = { text = "MIT" }
requires-python = "<4.0,>=3.9"
dependencies = [
    "gpiozero>=2.0,<3.0",
    "paho-mqtt<3.0.0,>=2.0.0",
    "schedule<2.0.0,>=1.1.0",
]

[dependency-groups]
dev = ["mypy>=1.15.0", "pytest>=8.3.4", "ruff>=0.9.7"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
